requirements.txt
fastapi
uvicorn
pandas
pydantic
psycopg2-binary
openpyxl
python-multipart
